Moves menu walker to its own file and adds class cleaning
- moves reverie_nav_walker to its own lib/nav.php file
- requires it in 4th position in functions.php
- adds additionnal menu-items class cleanup
- adds 'menu_type' key in $nav_args to enable class to reflect menu
type/menu location
